Add files via upload
Fixed: history.txt was not automatically created when it did not exist.
Added: renaming files if the title of the youtube video is in the structure "AUTHOR - TITLE ...", and removing parts such as (Official Video). This is useful when the video description does not contain mp3 metadata. The script also verifies that the author in AUTHOR - TITLE is the same than in the metadata of the video if it exists, so that when a music is reuploaded by someone else than the author it does not just take the youtube channel name as author in the mp3 file.